refactor(waves): satisfy sidebar debt guardrails
