+++ hooks/useWavesList.helpers.ts
@@ -1,6 +1,11 @@
 "use client";
 
 import { ApiWavesPinFilter } from "@/generated/models/ApiWavesPinFilter";
+import type { ApiWave } from "@/generated/models/ApiWave";
+import {
+  mapApiWaveOverviewToSidebarWave,
+  mapApiWaveToSidebarWave,
+} from "@/services/api/waves-v2-api";
 import type { SidebarWave } from "@/types/waves.types";
 
 export type SidebarDiscoverySection = "highly-rated" | "all";
@@ -13,6 +18,49 @@ export type SidebarWaveWithDiscoverySection = SidebarWave & {
 export const SIDEBAR_DISCOVERY_SECTION_HIGHLY_RATED: SidebarDiscoverySection =
   "highly-rated";
 export const SIDEBAR_DISCOVERY_SECTION_ALL: SidebarDiscoverySection = "all";
+
+interface ActiveSidebarContext {
+  readonly containerWave: SidebarWaveWithDiscoverySection | null;
+  readonly subwave: SidebarWave | null;
+}
+
+export const getActiveSidebarContext = (
+  activeWave: ApiWave | null
+): ActiveSidebarContext => {
+  if (activeWave === null) {
+    return { containerWave: null, subwave: null };
+  }
+
+  const activeSidebarWave = mapApiWaveToSidebarWave(activeWave);
+  if (activeSidebarWave.isDirectMessage) {
+    return { containerWave: null, subwave: null };
+  }
+
+  if (activeSidebarWave.parentWaveId === null) {
+    return {
+      containerWave: {
+        ...activeSidebarWave,
+        isInAllWaves: true,
+        sidebarSection: SIDEBAR_DISCOVERY_SECTION_ALL,
+      },
+      subwave: null,
+    };
+  }
+
+  if (!activeWave.parent_wave) {
+    return { containerWave: null, subwave: null };
+  }
+
+  return {
+    containerWave: {
+      ...mapApiWaveOverviewToSidebarWave(activeWave.parent_wave),
+      hasSubwaves: true,
+      isInAllWaves: true,
+      sidebarSection: SIDEBAR_DISCOVERY_SECTION_ALL,
+    },
+    subwave: activeSidebarWave,
+  };
+};
 
 const HIGHLY_RATED_WAVE_LIMIT = 10;
